FROMGIT: HID: google: whiskers: more robust tablet mode detection
The USB interface may get detected before the platform/EC one, so let's
note the state of the base (if we receive event) and use it to correctly
initialize the tablet mode switch state.

Also let's start the HID interface immediately when probing, this will
ensure that we correctly process "base folded" events that may be sent
as we initialize the base. Note that this requires us to add a remove()
function where we stop and close the hardware and switch the LED
registration away from devm interface as we need to make sure that we
destroy the LED instance before we stop the hardware.

Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 38e57f0693ad1b9a785b4f542ec019c5fdf91d57
 git://git.kernel.org/pub/scm/linux/kernel/git/hid/hid.git for-5.5/whiskers)

BUG=b:140906956
TEST=Boot krane, attach masterball

Change-Id: I5ca9367fe12e5c95f2fd5d6ff0850a1b164a9176
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1804020