BACKPORT: drm/i915: Don't advertise modes that exceed the max plane size
Modern platforms allow the transcoders hdisplay/vdisplay to exceed the
planes' max resolution. This has the nasty implication that modes on the
connectors' mode list may not be usable when the user asks for a
fullscreen plane. Seeing as that is the most common use case it seems
prudent to filter out modes that don't allow for fullscreen planes to
be enabled.

Let's do that in the connetor .mode_valid() hook so that normally
such modes are kept hidden but the user is still able to forcibly
specify such a mode if they know they don't need fullscreen planes.

This is in line with ealier policies regarding certain clock limits.
The idea is to prevent the casual user from encountering a mode that
would fail under typical conditions, but allow the expert user to
force things if they so wish.

Maybe in the future we should consider automagically using two
planes when one can't cover the entire screen? Wouldn't be a
great match for the current uapi with explicit planes though,
but I guess no worse than using two pipes (which we apparently
have to in the future anyway). Either that or we'd have to
teach userspace to do it for us.

v2: Fix icl+ max plane heigth (Manasi)

Cc: Manasi Navare <manasi.d.navare@intel.com>
Cc: Leho Kraav <leho@kraav.com>
Cc: Sean Paul <sean@poorly.run>
Cc: José Roberto de Souza <jose.souza@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Manasi Navare <manasi.d.navare@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20190918150707.32420-1-ville.syrjala@linux.intel.com
(cherry picked from commit 2d20411e25a3bf3d2914a2219f47ed48dc57aed5)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

Conflicts:
        drivers/gpu/drm/i915/display/intel_display.c
        drivers/gpu/drm/i915/display/intel_display.h
        drivers/gpu/drm/i915/display/intel_dp.c
        drivers/gpu/drm/i915/display/intel_hdmi.c
        These files did not exist in kernel v4.19.
        Backport the changes to i915/intel_display.c,
        i915/intel_display.h, i915/intel_dp.c and
        i915/intel_hdmi.c.

        drivers/gpu/drm/i915/display/intel_dsi.c
        i915/intel_dsi.c to support MIPI DSI did not
        exist in kernel 4.19. So ignore the change.

BUG=b:146523341, b:157098276
TEST=Connect monitor support HDMI 2.0 display to RT2141's HDMI port.
     Confirm DUT did not output 4k@60fps that exceed HDMI 1.4 limitation.

Change-Id: I020f47c638ff6484cd8f9a40c92238461384446e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2351951
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>