CHROMIUM: dt-bindings: drm/rockchip: dw_hdmi: Add nvmem entry for cpu ID
In order to decode the HDCP key that userspace provides the Rockchip
HDMI driver needs to be able to read the CPU ID out of nvmem.  Add for
this in the bindings.

This is a CHROMIUM patch because, as far as I can tell from looking at
Rockchip's github tree, only Chrome OS devices use the CPU ID like
this.  Other devices using Rockchip SoCs get the HDCP seed from their
downstream "vendor storage".  A more upstream friendly solution might
be to have userspace read the nvmem and provide it at the same time it
provides the key.

BUG=chromium:971474
TEST=None

Change-Id: I664c486185562819104d8f8a40ca34c7f5fdd429
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1814422
Reviewed-by: Sean Paul <seanpaul@chromium.org>