FIXUP: FROMLIST: drm/panel: support for boe, tv101wum-n53 wuxga dsi video mode panel
Boe,tv101wum-n53's need modify EDID parameter to fix the issue of abnormal
display of the top and bottom edges of the panel.

Signed-off-by: Jitao Shi <jitao.shi@mediatek.com>
(am from https://patchwork.kernel.org/patch/11150255/)
The difference between FROMLIST v5 and v6 is that v6 modifies the EDID parameter
of the panel, which is what this CL submits.

BUG=b:139731164, b:140545315
TEST=build and update kernel, panel can work well

Change-Id: Ibf00eeb22860dfbf4c70e02ee428c7b433419998
Signed-off-by: Hao He <hao.he@bitland.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1812439
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>