CHROMIUM: iwl7000: pcie: fix rb_allocator workqueue allocation
We don't handle failures in the rb_allocator workqueue allocation
correctly. To fix that, move the code earlier so the cleanup is
easier and we don't have to undo all the interrupt allocations in
this case.

Change-Id: I486160666c655cbd47e37c8f103b89ef685b6ae1
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 721173279fb7028769cb24b48f0de30feaf72402
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1974911
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>