CHROMIUM: ARM: dts: rockchip: Add nvmem entry for cpu ID to HDMI on rk3288
In order to load the HDCP key on rk3288 we need to use the CPU ID to
provide a seed.  Add a link from the HDMI driver to the CPU ID.

This is a CHROMIUM patch because, as far as I can tell from looking at
Rockchip's github tree, only Chrome OS devices use the CPU ID like
this.  Other devices using Rockchip SoCs get the HDCP seed from their
downstream "vendor storage".  A more upstream friendly solution might
be to have userspace read the nvmem and provide it at the same time it
provides the key.

BUG=chromium:971474
TEST=Use this to generate HDCP seed

Change-Id: I5deddea3258bae82997c53c576604574addca8f0
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1814423
Reviewed-by: Sean Paul <seanpaul@chromium.org>