CHROMIUM: drm/rockchip: dw_hdmi: Support Chromebook-style HDCP key from userspace
This patch adds code to the Rockchip HDMI driver to support loading
the HDCP key in a way that rk3288-veyron-mickey expects.
Specifically:
- rk3288-veyron-mickey expects to load the HDCP key (minus the seed)
  from userspace through sysfs.
- The HDCP key that userspace has access to needs to be combined with
  a "seed" which is generated by hashing the CPU ID.

It is expected that nobody other than rk3288-veyron-mickey works in
exactly this way.  Specifically from looking at Rockchip's github
tree:
- It seems that nobody else generates the seed by hashing the CPU ID.
- On other Rockchip devices userspace doesn't seem to be involved.  It
  seems that in Rockchip's downstream tree they use
  "rk_vendor_storage".

It is _possible_ that we could make this more generic by forcing
userspace to provide both the key and the seed.  On
rk3288-veyron-mickey userspace could be changed to read the CPU ID
directly, hash it, and provide it to the kernel.  If this code were to
make its way upstream and folks were OK w/ this sysfs interface that
would probably be the way to go.

For now we can make this a CHROMIUM patch.

BUG=chromium:971474
TEST=Use this to generate HDCP seed

Change-Id: I0cdfedbedf73a4199d5961ecce8e51a25b217c05
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1814892
Reviewed-by: Sean Paul <seanpaul@chromium.org>