CHROMIUM: virtwl: fix descriptor sizes for close, dmabuf_sync
The guest kernel was sending incorrectly-sized descriptors for the
VIRTIO_WL_CMD_VFD_CLOSE and VIRTIO_WL_CMD_VFD_DMABUF_SYNC messages; the
'out' descriptor is the one sent from the kernel to the device, and this
needs to be the full size of the message, but previously we were sending
a descriptor containing just the header as the 'out' descriptor.

This didn't make any difference in practice because the crosvm device
implementation did not check the descriptor sizes, but as part of an
effort to refactor the virtio devices to use a new descriptor
reader/writer abstraction, the sizes will be checked, uncovering this
issue.

BUG=chromium:966258
TEST=Run gnome-mahjongg under Crostini with CL:1815418 applied

Change-Id: Ica4c9ee6de71120d58aa9a1c90c1c9bb98580dd5
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1817137
Reviewed-by: Zach Reizner <zachr@chromium.org>