CHROMIUM: iwl7000: assume the driver_data is a trans_cfg, but allow full cfg
With the new concept of separating the trans-specific (trans_cfg) data
from the rest of the cfg, we will start mapping only the trans_cfg
part to the PCI device ID/subsystem device ID.  So we can assume that
the data passed to the probe function contains the trans_cfg, but
since the full cfg still contains the trans_cfg at the beginning, we
can allow a full cfg to be passed as well.  This makes it easier to
convert the existing tables one by one.

Change-Id: I7cd0208ef2471a47b106e4e13c408f0876fc7a7f
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 506071141820f1ae152e8fefd9e17cbd02f1cdb7