CHROMIUM: media: mtk-vcodec: stop using V4L2 structures to communicate with firmware
V4L2 structures are subject to change and should not be used to
communicate with other components, such as the firmware, even if said
firmware currently uses a similar structure. Replace the V4L2 structures
that were still used in the firmware-communicating parts of the driver
with equivalent ones and perform a deep copy.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

BUG=b:142687721
BUG=b:139640858
TEST=Kernel builds, boots and decodes videos on Kukui.

Change-Id: Icb3f05a0d3ae378a2a4d42d1d4276f97ff7684da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1866116
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>