UPSTREAM: Bluetooth: hci_qca: Don't vote for specific voltage
Devices with specific voltage requirements should not request voltage
from the driver, but instead rely on the system configuration to define
appropriate voltages for each rail.

This ensures that PMIC and board variations are accounted for, something
that the 0.1V range in the hci_qca driver currently tries to address.
But on the Lenovo Yoga C630 (with wcn3990) vddch0 is 3.1V, which means
the driver will fail to set the voltage.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit f2edd66e515b9944a7e516510a54959e5004181b)

BUG=b:142923859, b:145472775
TEST=bluetooth_AdapterQuickSanity

Change-Id: Ifc93a9a71db20a0afa1ba8bd5f7b4ab36609f4e9
Signed-off-by: Claire Chang <tientzu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2010453
Reviewed-by: Sean Paul <seanpaul@chromium.org>