CHROMIUM: kvm, kstaled: age VM pages in kstaled
This is based on the prod kernel implementation,
but has been simplified because we don't have some of the features
of the prodkernel and we can get away with less locking.

For each KVM instance we add a list of pagetable pages for
that instance.

We add a reference count to kvm_mmu_page to make sure that
the pagetable pages stay around while we are looking at them.
Additionally, we make sure that we wait an RCU grace period
before freeing them.

For data pages, we just get_page_unless_zero and verify that
it is the right page, using cmpxchg64 to make sure that the
PTE hasn't changed.

As a result, we can locklessly walk the list of KVM instances
and their respective lists of pagetable pages and collect/reset their
accessed bits.

TEST=Shrank RAM using /dev/ram0 and ran crostini.DiskIOPerf and
crostini.VimCompile with VM_BUG_ON enabled.  Observed that
kstaled_vm_background_hot is incremented.
BUG=b:123039911

Change-Id: I880ff3db6757f717485a98e5e4a3bd44fd2bef41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1898208
Tested-by: Salman Qazi <sqazi@google.com>
Reviewed-by: Yu Zhao <yuzhao@chromium.org>
Commit-Queue: Salman Qazi <sqazi@google.com>