CHROMIUM: iwl7000: yoyo: check for the domain on all TLV types during init
Now that we don't have dynamically changing domains anymore, we can
simply skip all the TLVs with domains that are not enabled.  To do so,
remove the checks from the functions that handle the TLVs when a
timepoint is reached to the top allocation function.

Change-Id: Ib9cf75bbde1f7a1e030d95066fbc016a8cecff30
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 376dbdf509e7fa4f2b3239d120523a77f593fadd