FIXUP: FROMLIST: thermal: mediatek: add another get_temp ops for thermal sensors
Fix thermal probing for mt8173.

Commit 5d2ff2ebaf78 ("FROMLIST: thermal: mediatek: add another get_temp
ops for thermal sensors") changes that thermal-zones needs to have
subnodes (eg. cpu_thermal, tzts..) amount equal to num_sensors + 1.
In mt8173 sensors are used in multiple thermal zones.

Bypass the failure that non cpu_thermal sensro is not find in thermal-zones
in dts, which is normal for mt8173, so prompt a warning here instead of
failing.

(replied in https://patchwork.kernel.org/patch/10938829/#23014505)

BUG=b:144254536, b:131870692
TEST=grep .  /sys/class/thermal/thermal_zone*/temp

Change-Id: I39043d246f179dc5c0b2d8ece8164e16c9f25197
Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1911839
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Michael Kao <michael.kao@mediatek.corp-partner.google.com>