BACKPORT: cpuidle: teo: Avoid using "early hits" incorrectly
If the current state with the maximum "early hits" metric in
teo_select() is also the one "matching" the expected idle duration,
it will be used as the candidate one for selection even if its
"misses" metric is greater than its "hits" metric, which is not
correct.

In that case, the candidate state should be shallower than the
current one and its "early hits" metric should be the maximum
among the idle states shallower than the current one.

To make that happen, modify teo_select() to save the index of
the state whose "early hits" metric is the maximum for the
range of states below the current one and go back to that state
if it turns out that the current one should be rejected.

Fixes: 159e48560f51 ("cpuidle: teo: Fix "early hits" handling for disabled idle states")
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 63f202e5edf161c2ccffa286a9a701e995427b15
 git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

Conflicts:
   drivers/cpuidle/governors/teo.c

BUG=chromium:1039194
TEST=Booting on Krane with cpuidle.governor=teo

Change-Id: I69e64e0601c3b66c41e9b0cb3536491da6c30911
Signed-off-by: Ikjoon Jang <ikjn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1994718
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>