CHROMIUM: iwl7000: pcie: detect the DMA bug and warn if it happens
Warn if the DMA bug is going to happen. We don't have a good
way of actually aborting in this case and we have workarounds
in place for the cases where it happens, but in order to not
be surprised add a safety-check and warn.

Change-Id: If5e419f6e7cff3a691b4434697e091b6db735652
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: e9cde6bd7bd4e3f5ba348916a2cd83a3b7daf589