HACK: CHROMIUM: media: mtk-mdp3: Compute plane address with data_offset
With V4L2 MEMORY DMABUF API, it is necessary to set a file descriptor on
each plane. The way to pass an offset with a DmaBuf is not defined in V4L2
specification. This is the case as it is needed to specify offsets because
Chromium creates one DmaBuf for the planes.

BUG=chromium:901264
BUG=b:143515730
TEST=camera preview & capture pass

Change-Id: I90d265bb385622d2f888b84233653bd9c94a601b
Signed-off-by: daoyuan huang <daoyuan.huang@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1919093
Commit-Queue: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>