FROMLIST: usb: overridable hub bInterval by device node
This patch enables hub device to override its own endpoint descriptor's
bInterval when the hub has a device node with "hub,interval" property.

When we know reducing autosuspend delay for built-in HIDs is better for
power saving, we can reduce it to the optimal value. But if a parent hub
has a long bInterval, mouse lags a lot from more frequent autosuspend.
So this enables overriding bInterval for a hard wired hub device only
when we know that reduces the power consumption.
(am from https://patchwork.kernel.org/patch/11257167/)

BUG=b:144246456
TEST=check masterball resume time

Signed-off-by: Ikjoon Jang <ikjn@chromium.org>
Acked-by: Alan Stern <stern@rowland.harvard.edu>

Change-Id: Ic3637a05e350a6598cd40ab16b79154e29071bde
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1916549
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Ikjoon Jang <ikjn@chromium.org>
Commit-Queue: Ikjoon Jang <ikjn@chromium.org>