BACKPORT: platform/chrome: cros-ec: Record event timestamp in the hard irq
To improve sensor timestamp precision, given EC and AP are in different
time domains, the AP needs to try to record the exact moment an event
was signalled to the AP by the EC as soon as possible after it happens.

First thing in the hard irq is the best place for this.

BUG=chromium:946178
TEST=tast run <ip> hardware.SensorRing with sensor ring enabled.
     iio-info with sensor ring disabled.

Acked-by: Jonathan Cameron <Jonathan.Cameron@kernel.org>
Acked-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
(cherry picked from commit 05a3c420eaa6857cb20afe7e3a3c39ed94a3b2c1)
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I8b7a386fc54c84c36dc997386afa83f810d22f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2228697