CHROMIUM: mm/swapfile.c: allow non-zram devices to be used for swap
We are only enabling disk based swap on specific devices we want to
use for experiments, so we want allow non-zram devices to be used.
Take advantage of the proc_poll_event activity counter, which indicates
that a swapon or swapoff has occurred, to prevent non-zram devices to
be used if there has already been a swapon event. This should allow non
zram block devices to be used in the initial swap.sh setup script on
reboot, but prevent them from being created after.

BUG=b:143378037
TEST=Used non-zram device for swap on hatch, called swapoff and then
tried to swapon different block device, which failed with Invalid argument.

Change-Id: I8a3cdb9a71fa67ceec5f13b007bac49f6114196a
Signed-off-by: Alexis Savery <asavery@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1924839
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>