FROMLIST: media: videobuf2-core: fix dprintk level
This dprintk is supposed to be informational, not an error. Set it to
the same level as the other messages related to memory allocations so
the kernel log does not get filled by messages during normal operation.

BUG=b:123551776
TEST=No kernel logs during normal decoder operation on Kukui.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
(am from https://patchwork.kernel.org/patch/11261579/)

Change-Id: I2d9492c4543f647985a23c0fb97421dfb465ca17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1936054
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>