CHROMIUM: media: mtk-vcodec: make flush buffer reusable for encoder
The flush buffer is a special buffer that tells the decoder driver to
send an empty CAPTURE frame to the client with V4L2_BUF_FLAG_LAST set.

We need similar functionality for the encoder ; however currently the
flush buffer depends on decoder-specific structures and thus cannot be
reused with the encoder.

Fix this by testing for this buffer by its VB2 address, and not through
a dedicated flag stored in a higher-level decoder structure. This also
allows us to remove said flag and simplify the code a bit.

Since the flush buffer should never be used in the stateless decoder,
also add safeguards to check against it.

BUG=b:123551776
BUG=b:144135251
TEST=tast video.DecodeAccel.h264* passing on Kukui.

Change-Id: Ic3264072ba72fee96c541a1d0104b3f4c8e7f198
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1936056
Reviewed-by: Tomasz Figa <tfiga@chromium.org>