CHROMIUM: iwl7000: move API version lookup to common code
The API version lookup is parsed from a TLV and should be in shared code
make make it reusable across all opmodes.

Also change the function names from mvm to fw, since this is not
mvm-specific anymore.

Additionally, since this function is not just a single line of code, it
shouldn't be inline.  Convert them to actual functions.

Change-Id: I5ede9cc25ee8de7b8d2b5c574f917a18971da734
Signed-off-by: Mordechay Goodstein <mordechay.goodstein@intel.com>
iwl7000-tree: 102034f948a1ed73399232910141abd8473568e6