BACKPORT: FROMLIST: drm/i915: Add DPCD quirk for AUO PSR2 panel
Currently on AUO PSR2 panel on Gen9 chromebook, we are observing
below issues:
(i) The display will show garbage after pressing sign
out icon in log in screen when wallpaper is one of Solid colors
& PSR2 is enabled
(ii) The characters of display is not clear when switch
OS mode to dev mode.

Before this patch, on this panel, we set idle frame count to 6
that is number of idle frames before entering PSR2 deep sleep
and the number of frames to enter into Selective update we set
to 1.

On this AUO panel, we suspect there is some DP synchronization
latency needed, due to which we are facing the above issues.

With current TCON of the AUO panel, DPCD reg
DP_SYNCHRONIZATION_LATENCY_IN_SINK (0x2009) offset is giving a
value of 0x0.

This patch sets idle frame count to 9 and frame count for selective
update to 9, after which we are not seeing the above mentioned issues.

Ideally this value needs to be corrected in TCON of the panel
since this value comes from DPCD reg 0x2009 offset and i915 driver
uses it. Working with AUO panel vendor to get this fixed in the
panel TCON. In the meantime fixing this as DPCD quirk in the kernel.

Signed-off-by: Gaurav K Singh <gaurav.k.singh@intel.com>
(am from https://patchwork.freedesktop.org/patch/344430/)

BUG=b:144747469,b:144747467
TEST= Boot Helios and make sure all issues mentioned above are resolved

Change-Id: Ia10634265a4549460d385a7d607b32a671212a10
Signed-off-by: Gaurav K Singh <gaurav.k.singh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1958165
Tested-by: Gaurav K Singh <gaurav.k.singh@intel.corp-partner.google.com>
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Gaurav K Singh <gaurav.k.singh@intel.corp-partner.google.com>
Commit-Queue: Evan Green <evgreen@chromium.org>