FIXUP: FROMLIST: arm64: dts: mt8183: add gce setting in display node
We need to include dt-bindings/gce/mt8183-gce.h before we use it.
The include file was previously included in commit 3b83dc6b1b7fd6d
("BACKPORT: FROMLIST: arm64: dts: add gce node for mt8183")
but this was removed as the change _itself_ does not require the include
file.

BUG=b:140360599
TEST=build and boot to shell

Change-Id: I42ebb6e1ddaa0cea3547a9ba9bd85f623c557c3d
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1969099
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Bibby Hsieh <bibby.hsieh@mediatek.com>