CHROMIUM: media: i2c: dw9768: Move vcm position smoothly when open/close
To decrease the noise appear when switching from Rear camera to front camera,
moving the lens position gradually in units of DW9768_CTRL_STEPS in dw9768_release,
to make the movements smoothly.

From vendor, DW9768 current operation time is 0.7*Tvib, about 8.4ms;
and the stabilization time is a delay added to prevent the motor from shaking,
approximately 20ms in the worst situation.

So it is recommended to wait current setting time for each step, and only wait for
the motor to stabilize after the last movement.
And DW9768_CTRL_STEPS is set to 16 to make vcm move enough smooth.

In addtion, reset 0x02 to BIT(0) that enable power-down bit when exiting.

BUG=b:146309921
TEST=verify pass

Change-Id: I03135b8dc510053294f9a09b13dd6184ed8526ee
Signed-off-by: Dongchun Zhu <dongchun.zhu@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1973850
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Hao He <hao.he@bitland.corp-partner.google.com>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>