CHROMIUM: iwl7000: fw: make pos static in iwl_sar_get_ewrd_table() loop
In the for loop where we are supposed to go through the entire table,
we are using a non-static local to keep the pos index.  This makes
each iteration start with 3, so we always access the first item on the
table.  Fix this by moving the variable outside of the loo so it
doesn't lose its value at every iteration.

Change-Id: Ifbbbc7b473eafce913c099719ba9bb33c4d08451
Reported-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: e09b9b3fd21ce5a5eb153ed2837826bbca883b0b