CHROMIUM: iwl7000: mvm: update the DTS measurement type
Till now, the driver asked the fw for a DTS measurement in automatic
mode. This triggered a flow in which the fw actively measured the
temperature.  This is not needed anymore, as the fw performs
measurements by itself, without the driver triggering them, and the
current cadence in which the fw performs such measurements is
sufficient.

In addition, in some time-sensitive scenarios, in which the driver asks
the fw for an active measurement twice in a short time (<100ms), the fw
asserts with code 0x20100801.

Change the DTS measurement to _WITHOUT_MEASURE instead, so the fw will
respond with the last measurement it has performed.

BUG=b:147005848

TEST=Follow the instructions in comment #1 of the bug report and make
    sure SYSASSERT 0x20100801 is not seen in the logs.

Change-Id: I53c9d76cc3ba946d7f93654ac7d6d0bd37b2ae31
Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
iwl7000-tree: 3378fdf716eba1ba19898757a795bcee18d0b8b8
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1992388
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>