CHROMIUM: rk3288_dmc: Avoid returning bogus busier channel
If both channels show 0 activity on read and write then we'll return a
bogus value in rk3288_dmc_get_busier_ch() and access some random
memory.  Let's stop doing that, shall we?  In the very bad case this
random memory will be out of range and crash us.

It's unclear why they would both be 0 but in any case this is a
nice/safe thing to do.

BUG=chromium:1035544
TEST=graphics_GLMark2.bvt

Change-Id: Ia702c677100ada671978f17208ac2217e268ace5
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1989865
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>