FROMLIST: platform/chrome: wilco_ec: Fix keyboard backlight probing
The EC on the Wilco platform responds with 0xFF to commands related to
the keyboard backlight on the absence of a keyboard backlight module.
This change allows the EC driver to continue loading even if the
backlight module is not present.

Signed-off-by: Daniel Campello <campello@chromium.org>

(am from https://lore.kernel.org/patchwork/patch/1176196)

BUG=b:146598606
TEST=Built and tested on sarien

Change-Id: Iedcdbae5a7ed79291b557882130e967f72168a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1988711
Tested-by: Daniel Campello <campello@chromium.org>
Tested-by: Lucien Kao <lucien_kao@compal.corp-partner.google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mathew King <mathewk@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>