CHROMIUM: [media] mtk-mdp3: Remove states for format checks
The mtk-mdp3 driver uses states to check if the formats have been set
on the capture and output when turning the streaming on, setting
controls or setting the selection rectangles.
Those states are reset when 0 buffers are requested like when checking
capabilities.
This patch removes all format checks and set one by default as queues in
V4L2 are expected to always have a format set.

https://linuxtv.org/downloads/v4l-dvb-apis/uapi/v4l/vidioc-streamon.html
https://linuxtv.org/downloads/v4l-dvb-apis/uapi/v4l/vidioc-g-ctrl.html
https://linuxtv.org/downloads/v4l-dvb-apis/uapi/v4l/vidioc-g-selection.html

This is a port of a similar patch in the mtk-mdp driver.

BUG=chromium:1020303
TEST=Video playback working on Kukui.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Change-Id: Ifb05a515a89cd095d13fdf2e9fdce7a86925bf81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1992964
Reviewed-by: Tomasz Figa <tfiga@chromium.org>