CHROMIUM: iwl7000: mvm: initialize iwl_dev_tx_power_cmd to zero
If the REDUCE_TX_POWER_CMD version is v4 or v5, we are not
initializing some values before sending to the FW, which causes SAR
not to work properly.  Solve this by initializing the struct in the
declaration.

Change-Id: I43cfd72d539c1287ccaaa454e95c673dac38214f
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 355eb6cff0bff6edf2f190c5df1a1911fbd7a651