CHROMIUM: media: mtk-vcodec: reset used flag when finishing buffer
The used flag helps detecting inconsistent states in the driver, but was
not reset when a buffer is removed from the queue due to a streamoff.
Set it to false in buf_finish to maintain a consistent state.

BUG=chromium:1023703
TEST=VideoDecoderTest.ResetMidStream does not complain about used buffers.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: If129ddf52820f098b8389367a5b5e34d796bf7e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2040983
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>