CHROMIUM: media: mtk-vcodec: complete pending requests during streamoff
Buffers that are queued during when a streamoff occurs are returned to
the user with an error set. However, their associated request (if any)
is never completed, which could lead userspace to wait endlessly for
them to complete.

BUG=chromium:1023703
TEST=VideoDecoderTest.ResetMidStream passes with VD decoder

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: Id71777b10030e6ae9e8323b4d13bf33f20898482
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2040984
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>