CHROMIUM: iwl7000: mvm: attempt to allocate smaller queues
We currently attempt to allocate queues that are 512 entries long,
but that requires 32 KiB memory, which may not be available, at
least not contiguously. If we fail to allocate, attempt to use a
smaller queue all the way down to 16 entries (which fit into a
single page).

Change-Id: I5059c410e628726cbce98d6311b690c632d00f97
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: c5e6d07792179b5d1382391ba679b1edac62260f