CHROMIUM: Handle powered down adapters in suspend notifier
If an adapter is powered down or in the process of being powered down,
it's possible for the suspend notifier to fail or corrupt the internal
state. In order to avoid that, always wait for power down to complete
and don't run the suspend notifier states if the adapter isn't powered
down.

BUG=b:150321586
TEST=Ran new bluetooth_AdapterSRSanity.sr_while_powering_down test and
     ran bluetooth_AdapterQuickSanity which originally caught the error.

Change-Id: I25950dda1fddca5af38ed11b4405ddc1e0037cf1
Signed-off-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2075993
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>