FIXUP: BACKPORT: FROMLIST: drm/bridge: add it6505 driver
Current code triggered a warning by kbuild test robot:
drivers/gpu/drm/bridge/ite-it6505.c:699 show_audio_m_count() error: uninitialized symbol 'aclk'.

This is because compiler can't reason that aclk is always initialized in
either the for loop or the if branch.

Change the code to always initialized aclk to 0 for the
SAMPLE_RATE_UNKNOWN case. Also take the chance to simplify the handling
of the unknown case, and use the ARRAY_SIZE(audio_sample_rate_map) as
bound instead of using a dummy item.

BUG=b:126007406
TEST=build

Change-Id: I59bbca746cda79bce3bc385c8fc9d7c6fc229db2
Signed-off-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2082737
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>