UPSTREAM: drm/dp_mst: Fix drm_dp_check_mstb_guid() return code
We actually expect this to return a 0 on success, or negative error code
on failure. In order to do that, we check whether or not we managed to
write the whole GUID and then return 0 if so, otherwise return a
negative error code. Also, let's add an error message here so it's a
little more obvious when this fails in the middle of a link address
probe.

This should fix issues with certain MST hubs seemingly stopping for no
reason in the middle of the link address probe process.

Fixes: cb897542c6d2 ("drm/dp_mst: Fix W=1 warnings")
Cc: Benjamin Gaignard <benjamin.gaignard@st.com>
Cc: Sean Paul <sean@poorly.run>
Cc: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Lyude Paul <lyude@redhat.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200306234923.547873-3-lyude@redhat.com
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 94b6ada401cbcf2e81f34ac7c5c4ca82154fd22e)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:146523341, b:157098276
TEST=Connect monitor support HDMI 2.0 display to RT2141's HDMI port.
     Confirm DUT did not output 4k@60fps that exceed HDMI 1.4 limitation.

Change-Id: I5d1b0b1f12ed0e7b835ac02721d14c6dc15b7e24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2333006
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>