CHROMIUM: low_mem: do not reserve the extra_free_kbytes
totalreserve_pages is the reserve of pages that are not available to
userspace allocations.  totalreserve_pages includes the high watermark
and extra_free_kbytes raises the high watermark.  Nullify the effect of
extra_free_kbytes by excluding it from the reserved pages to prevent
sending low memory notification too early.

It's porting the kernel 3.18 fix for caroline.  The 3.18 patch:
https://chromium-review.googlesource.com/1485627

BUG=b:150866007
TEST=check low memory notification is not sent when there are more than
400 MB of free memory on caroline.

Change-Id: I13d681e4c0d865b2f56c6387f9e8f8633aa3d582
Signed-off-by: Kuo-Hsin Yang <vovoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2107036
Reviewed-by: Alex Levin <levinale@chromium.org>