UPSTREAM: wireguard: receive: remove dead code from default packet type case
The situation in which we wind up hitting the default case here
indicates a major bug in earlier parsing code. It is not a usual thing
that should ever happen, which means a "friendly" message for it doesn't
make sense. Rather, replace this with a WARN_ON, just like we do earlier
in the file for a similar situation, so that somebody sends us a bug
report and we can fix it.

Reported-by: Fabian Freyer <fabianfreyer@radicallyopensecurity.com>
(cherry picked from commit 2b8765c52db24c0fbcc81bac9b5e8390f2c7d3c8)
Bug: 152722841
Change-Id: I40e109e5bac6541d309b588958ef0e09b3b9c19b
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>