CHROMIUM: iwl7000: mac80211: don't leave skb->next/prev pointing to stack
In beacon protection, don't leave skb->next/prev pointing to the
on-stack list, even if that's actually harmless since we don't use
them again afterwards.

While at it, check that the SKB on the list is still the same, as
that's required here. If not, the encryption (protection) code is
buggy.

Change-Id: Ia89fb79a0469d32137c9a04315a1d2dfc7b7d6f5
Fixes: 0a3a84360b37 ("mac80211: Beacon protection using the new BIGTK (AP)")
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: b1c5fd011070d6492a6460fa626e8aa9e69c64ed