UPSTREAM:Bluetooth: don't assume key size is 16 when the command fails
With this change, the encryption key size is not assumed to be 16 if the
read_encryption_key_size command fails for any reason.  This ensures
that if the controller fails the command for any reason that the
encryption key size isn't implicitely set to 16 and instead take a more
concervative posture to assume it is 0.

Signed-off-by: Alain Michaud <alainm@chromium.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 32b50729d91f0b59336851bf765b9423add00710)

BUG=b:152895475
TEST=See top commit for test details.

Change-Id: I1dbae5cb37620a5d4702cbf621d414996760dbc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2212875
Tested-by: Alain Michaud <alainm@chromium.org>
Commit-Queue: Alain Michaud <alainm@chromium.org>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>