UPSTREAM: platform/chrome: cros_ec_sensorhub: Add median filter
Events are timestamped in EC time space, their timestamps need to be
converted in host time space.

The assumption is the time delta between when the interrupt is sent
by the EC and when it is receive by the host is a [small] constant.
This is not always true, even with hard-wired interrupt. To mitigate
worst offenders, add a median filter to weed out bigger than expected
delays.

BUG=chromium:946178
TEST=tast run <ip> hardware.SensorRing with sensor ring enabled.
     iio-info with sensor ring disabled.

Acked-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Acked-by: Lee Jones <lee.jones@linaro.org>
Acked-by: Andy Shevchenko <andy.shevchenko@gmail.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
(cherry picked from commit 93fe48a585905675719835f8269258736de0948f)
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I0044a6b2336adde084d73b28099547b93113d992
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2228706
Reviewed-by: Sean Paul <seanpaul@chromium.org>