FROMLIST: platform/chrome: cros_ec_ishtp: skip old cros_ec responses
The ISHTP layer can give us old responses that we already gave up on. We
do not want to interpret these old responses as the current response we
are waiting for.

The cros_ish should only have one request in flight at a time. We send
the request and wait for the response from the ISH. If the ISH is too
slow to respond we give up on that request and we can send a new
request. The ISH may still send the response to the request that timed
out and without this we treat the old response as the response to the
current command. This is a condition that should not normally happen but
it has been observed with a bad ISH image. So add a token to the request
header which is copied into the response header when the ISH processes
the message to ensure that response is for the current request.

BUG=chromium:946178
TEST=compare with 5.6 cros_ec_ishtp.c.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Mathew King <mathewk@chromium.org>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
(am from https://lore.kernel.org/patchwork/patch/1221836/)
Change-Id: I99120c45a51945012b9f1c40bc54654965eca632
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2228715