FROMLIST: platform/chrome: cros_ec_ishtp: free ishtp buffer before sending event
Recycle the ISH buffer before notifying of a response or an event. Often
a new message is sent in response to an event and in high traffic
scenarios this can lead to exhausting all available buffers. We can
ensure we are using the fewest buffers possible by freeing buffers as
soon as they are used.

BUG=chromium:946178
TEST=compare with 5.6 cros_ec_ishtp.c

Signed-off-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Mathew King <mathewk@chromium.org>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
(am from https://lore.kernel.org/patchwork/patch/1221837/)

Change-Id: I54e25003a8b65f13e3c7788eb0440991bc38da6d
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2228716
Reviewed-by: Sean Paul <seanpaul@chromium.org>