Revert "FROMLIST: pinctrl: cherryview: Add quirk with custom translation of ACPI GPIO numbers"
The quirk breaks headset detection on Setzer board.
It is because Setzer has vendor HP instead of GOOGLE in the DMI.

Even if HP is added to the DMI matching field, there is an issue on
touchpad.

Revert that patch, and let the proper fix in vboot_reference fix the
crossystem issue.

This reverts commit 383d9bec92939d1d34c1097358590d75ddc87c2b.

BUG=b:143174998, b:153985719
TEST=check headphone detection works on setzer.

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: Id0c752b1a4d4c16a2b78f60f9a4112d35c5eea9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2150577
Commit-Queue: Alex Levin <levinale@chromium.org>
Reviewed-by: Alex Levin <levinale@chromium.org>