CHROMIUM: mm: clear young on pmd when possible in kstaled
We clear the accessed bit on a parent pmd entry after we've done so
on all its child pte entries so that next time we can skip all the
children after we see the accessed bit is not set on the parent.

Multiple VMAs can share one (non-huge) pmd entry. We don't want to
clear the accessed bit on such a parent while scanning one of the
VMAs because we would then mistakenly skip other VMAs. However, the
original condition check was too strict. For direct aging (rmap),
it's true we don't want to step on adjacent VMAs because we don't
know if they share a pmd entry with us or not. It's unsafe to access
vm_prev and vm_next in this case. For kstaled thread, however, we
can relax the check because we are scanning all VMAs. If multiple
VMAs share one pmd entry, we can safely clear the accessed bit when
we reach the last of them. (VMAs are sorted, i.e., vm_prev->vm_end
<= vm_next->vm_start.)

BUG=b:123039911
TEST=Ran platform.MemoryPressure and platform.MemoryStress
     Before this patch:
       38.49% lzo1x_1_do_compress
        9.07% kstaled_walk_pte
        5.02% page_vma_mapped_walk
        4.43% memset
        3.51% do_raw_spin_lock
        1.66% zram_bvec_rw
        1.38% shrink_page_list
        1.30% kstaled_reclaim_lru
        1.24% try_to_unmap_one
        1.24% obj_malloc

     After this patch:
       40.48% lzo1x_1_do_compress
        6.43% kstaled_walk_pte
        5.90% page_vma_mapped_walk
        4.67% memset
        3.33% do_raw_spin_lock
        1.78% zram_bvec_rw
        1.44% kstaled_reclaim_lru
        1.42% shrink_page_list
        1.37% try_to_unmap_one
        1.31% obj_malloc

     We spend about 2% less time in kstaled_walk_pte(). This was
     observed on both Octopus and Grunt.

Signed-off-by: Yu Zhao <yuzhao@google.com>
Change-Id: I69c09060a354f193356b629dded364f792e2bb38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2199810
Reviewed-by: Yu Zhao <yuzhao@chromium.org>
Tested-by: Yu Zhao <yuzhao@chromium.org>
Commit-Queue: Yu Zhao <yuzhao@chromium.org>